fix python test
