+++ dev/sparktestsupport/modules.py
@@ -565,6 +565,7 @@ pyspark_ml = Module(
         "pyspark.ml.tests.test_stat",
         "pyspark.ml.tests.test_training_summary",
         "pyspark.ml.tests.test_tuning",
+        "pyspark.ml.tests.pyspark.ml.tests.test_tuning_on_pin_thread_mode",
         "pyspark.ml.tests.test_util",
         "pyspark.ml.tests.test_wrapper",
     ],
